pyproject.toml
[tool.poetry]
name = "nao-grita-"
version = "0.1.0"
description = ""
authors = ["perebaj <perebaj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.8"
llama-index = "^0.12.13"
llama-index-llms-openai = "^0.3.14"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.14.1"
pytest = "^8.3.4"
pydantic = "^2.10.6"
scikit-learn = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
